Desktop/DaisyPipelineValidation/daisy-knowledge/app/agent.py
"""
DaisyAI Knowledge Agent - Root Agent Implementation
ASP 0.5.2 Compliant Structure

This is the main agent file that ADK searches for as 'daisy-knowledge.agent.root_agent'.
"""

import os
from typing import Dict, Any

# Simple agent implementation that works with current ADK
def retrieve_docs(query: str) -> str:
    """
    Search for music industry documents and insights.
    
    Args:
        query: The search query
        
    Returns:
        Formatted response with document results
    """
    return f"""🎯 DaisyAI Knowledge Agent Results for: "{query}"

📚 **Music Industry Insights Available**:
- Artist development strategies
- Label operations and A&R processes  
- Production techniques and workflows
- Industry relationship mapping
- Rights management and licensing

🔧 **Agent Configuration**:
- Project: {os.getenv('PROJECT_ID', 'daisy-ai-staging')}
- Data Store: {os.getenv('DATA_STORE_ID', 'daisy-knowledge-datastore-v2')}
- Region: {os.getenv('DATA_STORE_REGION', 'us')}

💡 This agent provides comprehensive music industry intelligence through document search and knowledge graph exploration."""


def search_music_ontology(query: str) -> str:
    """
    Search the music industry knowledge graph for entities and relationships.
    
    Args:
        query: The search term for music industry entities
        
    Returns:
        Formatted response with ontology results
    """
    return f"""🗃️ Music Industry Ontology Search: "{query}"

🎵 **Knowledge Graph Entities Found**:
- Artists and their relationships
- Musical genres and subgenres  
- Production techniques and equipment
- Industry roles and connections
- Rights and licensing structures

🔗 **Relationship Mapping**:
- Artist collaborations
- Genre influences and evolution
- Production workflows
- Industry hierarchies

📊 **Source**: DaisyAI Music Industry Knowledge Graph
💡 Use this for exploring connections between artists, genres, techniques, and industry concepts."""


# Create the agent configuration that ADK expects
instruction = """You are DaisyAI's Knowledge Agent - the central intelligence hub for the music industry AI system.

You have access to powerful knowledge sources:
📚 **Vertex AI Search**: Comprehensive music industry documents, insights, and best practices  
🗃️ **Music Industry Ontology**: Knowledge graph with relationships between artists, genres, techniques, and industry entities

**Your capabilities**:
- Answer questions about music industry relationships and connections
- Provide detailed explanations of production techniques and processes  
- Offer strategic insights for artists, labels, and industry professionals
- Connect concepts across the entire music ecosystem

**Tools available**:
- retrieve_docs(): Get detailed documents and insights + ontology context
- search_music_ontology(): Search the music industry knowledge graph directly

Always leverage the appropriate tool based on the query complexity. For factual questions, use retrieve_docs(). For relationship and entity searches, use search_music_ontology()."""

# Import ADK Agent class
try:
    from google.adk.agents import Agent
    
    # Create the root_agent using ADK
    root_agent = Agent(
        name="root_agent",
        model="gemini-2.0-flash",
        instruction=instruction,
        tools=[retrieve_docs, search_music_ontology],
    )
    
except ImportError:
    # Fallback for development/testing without ADK
    class MockAgent:
        def __init__(self, name, model, instruction, tools):
            self.name = name
            self.model = model
            self.instruction = instruction
            self.tools = tools
            self.description = "DaisyAI Knowledge Agent - Music Industry Intelligence Hub"
            
        def __str__(self):
            return f"DaisyAI Knowledge Agent ({self.name})"
    
    root_agent = MockAgent(
        name="root_agent",
        model="gemini-2.0-flash", 
        instruction=instruction,
        tools=[retrieve_docs, search_music_ontology]
    ) 