Desktop/DaisyPipelineValidation/daisy-knowledge/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daisy-knowledge"
version = "1.0.0"
description = "DaisyAI Knowledge Agent - Central intelligence hub for the music industry AI system"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "DaisyAI Team", email = "info@daisyai.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["ai", "music", "industry", "knowledge", "agent"]
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.5.2",
    "vertexai>=1.38.0",
    "google-cloud-aiplatform>=1.38.0",
    "langchain-google-vertexai>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true 